refactor: coroutines를 사용한 비동기 및 병렬 처리
