+++ batch/src/main/kotlin/dnd11th/blooming/batch/weather/WeatherCareMessageJobConfig.kt
@@ -9,7 +9,6 @@ import org.springframework.batch.core.job.builder.JobBuilder
 import org.springframework.batch.core.launch.support.RunIdIncrementer
 import org.springframework.batch.core.repository.JobRepository
 import org.springframework.batch.core.step.builder.StepBuilder
-import org.springframework.batch.item.ItemProcessor
 import org.springframework.batch.item.ItemReader
 import org.springframework.batch.item.ItemWriter
 import org.springframework.context.annotation.Bean
@@ -39,13 +38,11 @@ class WeatherCareMessageJobConfig {
         jobRepository: JobRepository,
         transactionManager: PlatformTransactionManager,
         weatherCareMessageItemReader: ItemReader<Region>,
-        weatherCareMessageItemProcessor: ItemProcessor<Region, WeatherCareMessage>,
-        weatherCareMessageItemWriter: ItemWriter<WeatherCareMessage>,
+        weatherCareMessageItemWriter: ItemWriter<Region>,
     ): Step {
         return StepBuilder("weatherBatchStep", jobRepository)
-            .chunk<Region, WeatherCareMessage>(CHUNK_SIZE, transactionManager)
+            .chunk<Region, Region>(CHUNK_SIZE, transactionManager)
             .reader(weatherCareMessageItemReader)
-            .processor(weatherCareMessageItemProcessor)
             .writer(weatherCareMessageItemWriter)
             .build()
     }
