+++ client/src/main/kotlin/dnd11th/blooming/client/config/FeignClientConfig.kt
@@ -7,8 +7,4 @@ import org.springframework.context.annotation.Configuration
 @Configuration
 @EnableFeignClients(basePackages = ["dnd11th.blooming"])
 class FeignClientConfig {
-    @Bean
-    fun errorDecoder(): CustomErrorDecoder {
-        return CustomErrorDecoder()
-    }
 }
