+++ client/src/main/kotlin/dnd11th/blooming/client/weather/WeatherInfoClient.kt
@@ -1,6 +1,5 @@
 package dnd11th.blooming.client.weather
 
-import dnd11th.blooming.client.config.FeignClientConfig
 import org.springframework.cloud.openfeign.FeignClient
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestParam
@@ -8,7 +7,6 @@ import org.springframework.web.bind.annotation.RequestParam
 @FeignClient(
     name = "WeatherInfoClient",
     url = "http://apis.data.go.kr/1360000/VilageFcstInfoService_2.0/getVilageFcst",
-    configuration = [FeignClientConfig::class]
 )
 interface WeatherInfoClient {
     companion object {
